enhancement(admin): move widget visibility helper out of layout utils
Keep homepage widget display rules separate from sizing and positioning helpers so future visibility filters have a clearer home.
