+++ packages/core/admin/admin/src/utils/tests/widgetLayout.test.ts
@@ -10,8 +10,6 @@ import {
   isLastWidgetInRow,
   canResizeBetweenWidgets,
   createDefaultWidgetWidths,
-  hideDeployNowWidgetInProduction,
-  DEPLOY_NOW_WIDGET_UID,
 } from '../widgetLayout';
 
 import type { WidgetWithUID } from '../../core/apis/Widgets';
@@ -199,29 +197,5 @@ describe('createDefaultWidgetWidths', () => {
       'widget-2': 6,
       'widget-3': 12,
     });
-  });
-});
-
-describe('hideDeployNowWidgetInProduction', () => {
-  const deployNowWidget = createMockWidget(DEPLOY_NOW_WIDGET_UID, 'Deploy');
-  const widgetsWithDeploy = [...mockWidgets, deployNowWidget];
-
-  it('removes the deploy-now widget in production', () => {
-    const result = hideDeployNowWidgetInProduction(widgetsWithDeploy, 'production');
-
-    expect(result).toEqual(mockWidgets);
-    expect(result.map((widget) => widget.uid)).not.toContain(DEPLOY_NOW_WIDGET_UID);
-  });
-
-  it('keeps the deploy-now widget outside production', () => {
-    expect(hideDeployNowWidgetInProduction(widgetsWithDeploy, 'development')).toEqual(
-      widgetsWithDeploy
-    );
-  });
-
-  it('keeps the deploy-now widget when the environment is unknown', () => {
-    expect(hideDeployNowWidgetInProduction(widgetsWithDeploy, undefined)).toEqual(
-      widgetsWithDeploy
-    );
   });
 });
